+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Tämä dokumentti sisältää ytimekkään kuvauksen mitä muutoksia ohjelman ve
 
 * Bugikorjaus ranking pisteytyksiin, ei osannut laskea oikein kun ranking ykkönen, kakkonen tai kolmonen ei oo kisoissa mukana
 * Bugikorjaus tulosten näyttämiseen kun kisa on kesken. Saattoi näyttää 3.sijan väärin
+* Korjattu jaetut sijat pois tulosten alkupäästä kun kilpailu on vielä kesken
 
 ## Versio 1.0.0.18		(julkaistu 29.1.2025)